backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML
openai==1.3.5
anthropic==0.7.1
langchain==0.0.340
sentence-transformers==2.2.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# HTTP Requests
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.1

# Data Sources
PyGithub==2.1.1
feedparser==6.0.10

# Environment Management
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1

# Task Queue
celery==5.3.4

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
